scripts/nba/requirements.txt
nba_api==1.4.1
supabase==2.6.0
pandas==2.2.2
python-dateutil==2.9.0
tenacity==9.0.0
